+++ samples/callback.js
@@ -1,7 +1,11 @@
 
-should = require('should');
-generate = require('../lib');
+const generate = require('../lib')
+const assert = require('assert')
 
-generate({seed: 1, columns: 2, length: 2}, function(err, output){
-  output.should.eql('OMH,ONKCHhJmjadoA\nD,GeACHiN');
-});
+generate({
+  seed: 1,
+  columns: 2,
+  length: 2
+}, function(err, output){
+  assert.strictEqual(output, 'OMH,ONKCHhJmjadoA\nD,GeACHiN')
+})